perf(database): remove related animes caching due un usage and optionality
Additionally change internal ids data type to Long for full compatibility with Room

+++ app/src/main/java/live/shirabox/shirabox/ui/activity/player/PlayerViewModel.kt
@@ -20,7 +20,7 @@ import live.shirabox.shirabox.db.AppDatabase
 
 class PlayerViewModel(
     context: Context,
-    val contentUid: Int,
+    val contentUid: Long,
     val contentName: String,
     val episode: Int,
     val startIndex: Int,
